[Add]: ATG API endpoint
Since the sql server is currently not active at Jaco. I have mocked
the data that will be returned by a query in this API endpoint.

It is composed of an array of json objects.
Each json object is composed with a store number, a fuel grade,
a gas level and a date.

In order to request this data from the frontend, you will need to request
this api endpoint /api/ATG

and this will return the array of json values.